backend/requirements.txt
accelerate==1.1.1
apturl==0.5.2
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
diffusers==0.31.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.16.1
Flask==2.0.1
fonttools==4.54.1
fsspec==2024.10.0
future==0.18.2
gdown==5.2.0
gpg==1.16.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.26.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.43.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.9.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
paramiko==2.9.3
peft==0.13.2
pexpect==4.8.0
pillow==11.0.0
ply==3.11
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynndescent==0.5.13
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2024.11.6
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
scour==0.38.2
seaborn==0.13.2
SecretStorage==3.3.1
sentencepiece==0.2.0
sgt-launcher==0.2.7
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sports @ git+https://github.com/roboflow/sports.git@06053616f1f8a8ae1fa936eb00dcdc2e4f888bb1
supervision==0.24.0
sympy==1.13.1
systemd-python==234
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.2
triton==3.1.0
ttystatus==0.38
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.28
ultralytics-thop==2.0.10
umap-learn==0.5.7
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.0.2
xcffib==0.11.1
xdg==5
xkit==0.0.0
zipp==1.0.0
